Refactor/packager (#11)
`Transformer` and `Loader` have already been refactored into classes, so
this moves `Packager` to also become a class. I realized later on that
the class aspect didn't actually gain _too_ much here, but the bigger
benefit was refactoring out the Evaluator functions so that they can be
tested, and to limit the scope of `packager.js` to _just_ the packaging
and not also including all of the evaluation logic as well.

More refactoring will definitely happen later (`processCode` is too long
for my taste, etc.), but this is a big step to make it simpler to
implement new things like `Exclude` from #1 and such.

+++ packages/ts-docs-loader/test/webpack/compiler.js
@@ -56,7 +56,9 @@ export default function compiler(entrypoint, options = {}) {
  * @returns {Record<string, string>}
  */
 export function createFixtures(files) {
-  fs.rmdirSync(FIXTURES_DIR, {recursive: true});
+  if (fs.existsSync(FIXTURES_DIR)) {
+    fs.rmdirSync(FIXTURES_DIR, {recursive: true});
+  }
   fs.mkdirSync(FIXTURES_DIR, {recursive: true});
 
   /** @type {Record<string, string>} */
